requirements.txt
atari-py==0.2.6
attrs==19.3.0
box2d-py==2.3.8
cloudpickle==1.2.1
colorama==0.4.3
filelock==3.0.12
gtimer==1.0.0b5
gym~=0.17.2
matplotlib==3.1.2
more-itertools==8.0.2
mpi4py==3.0.3
mujoco-py==2.0.2.5
multidict==4.7.5
munch==2.5.0
numpy~=1.18.5
pybullet==2.6.1
PyYAML~=5.3.1
termcolor==1.1.0
torch~=1.6.0
scipy~=1.5.1
tqdm~=4.46.1
